requirements.txt
# ============================================
# 핵심 프레임워크
# ============================================
streamlit>=1.30.0

# ============================================
# 데이터 처리 & 성능
# ============================================
pandas>=2.1.0
numpy>=1.24.0
polars>=0.20.0              # 빠른 데이터프레임 처리
pyarrow>=14.0.0             # Apache Arrow (고성능 데이터 처리)

# ============================================
# 데이터베이스 & ORM
# ============================================
sqlalchemy>=2.0.0           # SQL ORM
alembic>=1.13.0             # 데이터베이스 마이그레이션
sqlmodel>=0.0.14            # Pydantic + SQLAlchemy 통합

# ============================================
# API & 네트워크
# ============================================
requests>=2.31.0
httpx>=0.26.0               # 비동기 HTTP 클라이언트
aiohttp>=3.9.0              # 비동기 HTTP 서버/클라이언트
python-dotenv>=1.0.0

# ============================================
# 데이터 검증 & 직렬화
# ============================================
pydantic>=2.5.0
pydantic-settings>=2.1.0
marshmallow>=3.20.0         # 데이터 직렬화/검증

# ============================================
# 파일 처리 & 문서 관리
# ============================================
openpyxl>=3.1.0
python-magic>=0.4.27        # 파일 타입 감지 (선택사항)
python-pptx>=0.6.23         # PowerPoint 처리
PyPDF2>=3.0.0               # PDF 텍스트 추출
streamlit-pdf-viewer>=0.0.1 # PDF 뷰어
xlsxwriter>=3.1.0           # Excel 생성

# ============================================
# UI 컴포넌트 & 디자인
# ============================================
streamlit-aggrid>=0.3.4
plotly>=5.18.0
streamlit-option-menu>=0.3.6    # 네비게이션 메뉴
streamlit-extras>=0.3.6         # 유용한 추가 컴포넌트
streamlit-lottie>=0.0.5         # Lottie 애니메이션
streamlit-elements>=0.1.0        # Material UI 컴포넌트
streamlit-card>=0.0.4           # 카드 컴포넌트
streamlit-calendar>=0.6.0       # 캘린더 위젯
streamlit-shadcn-ui>=0.1.0      # 모던한 UI 컴포넌트 (버전 완화)
streamlit-antd-components>=0.1.0 # 깔끔한 메뉴/탭 (버전 완화)

# ============================================
# 차트 & 시각화
# ============================================
matplotlib>=3.8.0
seaborn>=0.13.0
altair>=5.2.0

# ============================================
# Git & 버전 관리
# ============================================
GitPython>=3.1.40           # Python에서 Git 조작
watchdog>=3.0.0             # 파일 변경 감지

# ============================================
# 작업 스케줄링 & 자동화
# ============================================
schedule>=1.2.0             # 작업 스케줄러
apscheduler>=3.10.0         # 고급 스케줄러

# ============================================
# 캐싱 & 성능
# ============================================
diskcache>=5.6.0            # 디스크 기반 캐시
cachetools>=5.3.0           # 메모리 캐시 유틸리티

# ============================================
# 로깅 & 모니터링
# ============================================
loguru>=0.7.2
rich>=13.7.0                # 터미널 출력 미화
colorama>=0.4.6             # 색상 출력 (Windows 호환)

# ============================================
# 유틸리티
# ============================================
python-dateutil>=2.8.0
tqdm>=4.66.0
click>=8.1.0                # CLI 도구
typer>=0.9.0                # 현대적 CLI 프레임워크

# ============================================
# 테스트 & 품질 (개발용)
# ============================================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0      # 비동기 테스트
black>=23.12.0
ruff>=0.1.0
mypy>=1.7.0

# ============================================
# 배포 & 프로덕션
# ============================================
gunicorn>=21.2.0            # WSGI 서버 (Linux/Mac)
uvicorn>=0.25.0             # ASGI 서버
python-multipart>=0.0.6     # 파일 업로드 지원

# ============================================
# 보안
# ============================================
cryptography>=41.0.0        # 암호화
python-jose>=3.3.0          # JWT 토큰
passlib>=1.7.4              # 비밀번호 해싱
